+++ CHANGELOG.md
@@ -29,7 +29,7 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [v6.6.4](https://github.com/asfadmin/Discovery-asf_search/compare/v6.6.3...v6.6.4)
 ### Added
 - Adds new `dataset` keyword to `search()` as an alternative to `platform`. Allows users to get results from multiple platforms at once in a single page.
-- adds `frameNumber` properties support for new `Sentinel-1 Interferogram (BETA)` products
+- adds `frameNumber` properties support for new `Sentinel-1 Interferogram` products
 
 ### Changed
 - Changes `CMR_FORMAT_EXT` constant from `umm_json_v1_4` to `umm_json`, umm returned from CMR will now be in latest umm format by default