+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Adds `operaBurstID` keyword to `search()`
 - Adds OPERA-S1 param `operaBurstID` to `ASFProduct.properties`, and adds Opera product urls to `additionalUrls`
 - OPERA-S1 RTC product `polarization` now shows both polarizations as list
+- adds `frameNumber` properties support for new `Sentinel-1 Interferogram` products
 
 ### Changed
 - Changes `CMR_FORMAT_EXT` constant from `umm_json_v1_4` to `umm_json`, umm returned from CMR will now be in latest umm format by default